start of tradeController, with slight type change

+++ Libraries/Core/Models/Eft/Trade/ProcessRagfairTradeRequestData.cs
@@ -27,5 +27,5 @@ public record ItemRequest
     public string? Id { get; set; }
 
     [JsonPropertyName("count")]
-    public int? Count { get; set; }
+    public double? Count { get; set; }
 }
